requirements.txt
oss2==2.7.0
pillow>=8.0.0
pytest-django==3.9.0
pytest-mock==3.2.0